+++ CHANGELOG.md
@@ -7,20 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2024-06-23
+
+### Added
+- ``MessageHandler`` classes for character, join and ready requests, using *Chain of Responsibility*
+- ID Generator
+- Skeleton for ``Server`` class (#16)
+- Lobby system & connection classes using the mediator design pattern
+
+### Removed
+- ``GameController`` class
+
 ## [0.0.1] - 2024-06-16
 
 ### Added
 - JSON protocol for Server-Client communication with sequence diagram (#3, #4).
-- README.md.
+- ``README.md``.
 - Assignment of tasks for each student (#6).
 - UML diagram.
 - Information about the GUI (#2).
 - Information about the game logic (#11).
 - Packages and the relevant classes based on MVC pattern.
-- Observable and Observer in util package.
-- build.gradle configuration.
+- ``Observable`` and ``Observer`` in ``util`` package.
+- ``build.gradle`` configuration.
 - UI Classes.
-- Javadoc.
+- Javadocs.
 
 ### Changed